plugins/org.eclipse.eef.core/src/org/eclipse/eef/core/internal/controllers/EEFImageViewerController.java
/*******************************************************************************
 * Copyright (c) 2016 Obeo.
 * All rights reserved. This program and the accompanying materials
 * are made available under the terms of the Eclipse Public License v1.0
 * which accompanies this distribution, and is available at
 * http://www.eclipse.org/legal/epl-v10.html
 *
 * Contributors:
 *    Obeo - initial API and implementation
 *******************************************************************************/
package org.eclipse.eef.core.internal.controllers;

import java.util.HashMap;
import java.util.Map;

import org.eclipse.core.runtime.IStatus;
import org.eclipse.eef.EEFImageViewerDescription;
import org.eclipse.eef.EEFWidgetDescription;
import org.eclipse.eef.EefPackage;
import org.eclipse.eef.core.api.EEFExpressionUtils;
import org.eclipse.eef.core.api.EditingContextAdapter;
import org.eclipse.eef.core.api.controllers.AbstractEEFWidgetController;
import org.eclipse.eef.core.api.controllers.IConsumer;
import org.eclipse.eef.core.api.controllers.IEEFImageViewerController;
import org.eclipse.eef.core.api.utils.EvalFactory;
import org.eclipse.emf.ecore.EAttribute;
import org.eclipse.sirius.common.interpreter.api.IInterpreter;
import org.eclipse.sirius.common.interpreter.api.IVariableManager;

/**
 * This class will be used in order to manage the behavior of the image viewer.
 *
 * @author arichard
 */
public class EEFImageViewerController extends AbstractEEFWidgetController implements IEEFImageViewerController {
	/**
	 * The description.
	 */
	private EEFImageViewerDescription description;

	/**
	 * The editing context adapter.
	 */
	private EditingContextAdapter contextAdapter;

	/**
	 * The consumer of a new path of the image.
	 */
	private IConsumer<String> newPathConsumer;

	/**
	 * The constructor.
	 *
	 * @param description
	 *            The description
	 * @param variableManager
	 *            The variable manager
	 * @param interpreter
	 *            The interpreter
	 * @param contextAdapter
	 *            The editing context adapter
	 */
	public EEFImageViewerController(EEFImageViewerDescription description, IVariableManager variableManager, IInterpreter interpreter,
			EditingContextAdapter contextAdapter) {
		super(variableManager, interpreter);
		this.description = description;
		this.contextAdapter = contextAdapter;
	}

	@Override
	public IStatus updatePath(final String path) {
		return this.contextAdapter.performModelChange(() -> {
			String editExpression = EEFImageViewerController.this.description.getEditExpression();
			EAttribute eAttribute = EefPackage.Literals.EEF_IMAGE_VIEWER_DESCRIPTION__EDIT_EXPRESSION;

			Map<String, Object> variables = new HashMap<String, Object>();
			variables.putAll(EEFImageViewerController.this.variableManager.getVariables());
			variables.put(EEFExpressionUtils.EEFImageViewer.NEW_PATH, path);

			EvalFactory.of(EEFImageViewerController.this.interpreter, variables).logIfBlank(eAttribute).call(editExpression);
		});
	}

	/**
	 * {@inheritDoc}
	 *
	 * @see org.eclipse.eef.core.api.controllers.AbstractEEFWidgetController#refresh()
	 */
	@Override
	public void refresh() {
		super.refresh();

		String valueExpression = this.description.getPathExpression();
		this.newEval().logIfInvalidType(String.class).call(valueExpression, this.newPathConsumer);
	}

	/**
	 * {@inheritDoc}
	 *
	 * @see org.eclipse.eef.core.api.controllers.IEEFImageViewerController#onPathValue(org.eclipse.eef.core.api.controllers.IConsumer)
	 */
	@Override
	public void onNewPath(IConsumer<String> consumer) {
		this.newPathConsumer = consumer;
	}

	/**
	 * {@inheritDoc}
	 *
	 * @see org.eclipse.eef.core.api.controllers.IEEFImageViewerController#removeNewPathConsumer()
	 */
	@Override
	public void removeNewPathConsumer() {
		this.newPathConsumer = null;
	}

	/**
	 * {@inheritDoc}
	 *
	 * @see org.eclipse.eef.core.api.controllers.AbstractEEFWidgetController#getDescription()
	 */
	@Override
	protected EEFWidgetDescription getDescription() {
		return this.description;
	}

}
